ia_consultant/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langchain-google-genai
faiss-cpu
pydantic
python-dotenv
unstructured
python-docx
jq
gunicorn 